Refactor utils/print to consoleprint package
Renamed utils/print.go and utils/print_test.go to consoleprint/consoleprint.go and consoleprint/consoleprint_test.go, respectively. Updated package references to reflect the new structure.
